Save default parameter annotation to join post-inference
Summary:
This diff implements a suggestion for handling type inference that accounts for default parameter values based on the discussion in D34161077 (https://github.com/facebook/pyre-check/commit/bc0ea7cd46402fe4c5ee5405a048a8ab676ad081).

 ---

First, some background context on why special handling is necessary:

The basic structure of inference does a forward pass, which joins types that are assigned to each variable. Then the backwards pass performs meets of the types the variable is used as.

If we simply set the type of the parameter's default value as the starting annotation for the variable, we run into a problem:

```
def foo(x = None):
   takes_optional_int(x)
```

is actually quite different from

```
def foo(x):
    x = None
    takes_optional_int(x)
```

In the first example, `x` can possibly be `None` when the function call happens, but it could also be the value of whatever non-default input came in. In the second example, `x` can only possibly be `None` by the time the function call happens.

The refinement happening when we meet the type of `x` with the usage as `Optional[int]` therefore has no bearing on what the possible annotation of `x` is at the input in the second case, and our type inference correctly is not sure of the answer.

However, this is not sufficient behavior for the first example, where the refinement happening on the call does in fact restrict what the possible input values for `x` into the function are. We should be inferring an annotation of `Optional[int]`.

 ---

The current state of the world in inference is even messier - due to deeper issues in how `meet` refines types in the backwards pass, the logic used to simply replace the inferred type of the variable with the type of how it's used.

When thinking about how to implement this join-after-the-fact structure for default parameter types, I wanted to actually store the information on the state itself and not some special-cased extra pass on the define node to collect this information separately and apply it to the state produced from inference.

It turns out this kind of "save a type to join it in later" is exactly the behavior we'd need to robustly support backwards meets in a world where we're attempting to infer the overall type annotation of a variable rather than the annotation at any given point. Hence this diff introduces the idea of a "snapshot_resolution" which is extended in D34305388 to fix deeper issues with infer.

Reviewed By: dkgi

Differential Revision: D34292909

fbshipit-source-id: 117d16cf6d92621723db717255357a80848cb056