Not ignore any implicit targets from `str(e)` in f strings
Summary: Consider evaluating expression `e` when `e` is inside an f string. Implicit targets are generated from the implicit call to `str(e)`. Currently we ignore some of the targets such as `int.__str__`, which is however incorrect, because for example if a user-defined class inherits `int` and overrides method `__str__`, then the behavior of `int.__str__` may not be the default tito. That is, it may create sources or contain sinks.

Reviewed By: arthaud

Differential Revision: D34403575

fbshipit-source-id: c231739de0a20990cce5717510df941b8c72ebfa