Properly resolve implicit calls from `str(e)` to `e.__str__()` when having multiple expressions at the same locations
Summary:
There is a very subtle bug due to `a + b` being translated into `a.__add__(b)` where `a.__add__` has the same location of `a`.
This leads to having multiple expressions with the same locations, which we handle using a `Compound` type in the call graph, with an expression identifier.
The problem is when redirecting calls from `str(e)` to `e.__str__()`, the expression identifier changes from `str` to `__str__`.
To prevent these problems, we should redirect calls before call resolution and registration.

Reviewed By: dkgi

Differential Revision: D34468241

fbshipit-source-id: f7a22b4bc5d1afab18205e0aa7d5cc4d1e2f88a9