Log time taken to create lookup for a module.
Summary:
Problem: When responding to a definition query, Pyre currently creates a lookup table for all names in a module. This means the go-to-def query takes time proportional to the size of a file; e.g., ~700 ms for a 1800 LOC file such as `persistent.py`. Pretty much all of this time is in creating a lookup table for the module.

Another problem is that we currently couple the lookup for definitions and the lookup for resolved types. This means we are essentially doubling the time taken for both the `go-to-def` query and the `types` query.

Solution:
+ When it comes to go-to-def for a location, we can search for the precise name that covers that location. That way, we will only lookup the definition for that one name instead of all names in a module.
+ This means we can stop computing the definitions lookup for the `types(...)` query, which is used for hover. Will optimize hover further based on my experiments with `go-to-def`.

## Stack plan

+ This diff just adds perf logging. I'm not writing it to the perf table since we are going to change the strategy from module-level lookup to just a single expression lookup.
+ Next, add code to find the name covering a location.
+ Finally, look up the definition for that name.

Reviewed By: grievejia

Differential Revision: D34520606

fbshipit-source-id: 2f4b19558cf2c150a010d4f5166fa51e5d2828d3