Implement query for `location_to_definition`
Summary:
Problem: We'd set up the request and response variants but had not implemented the actual response.

We build a location-lookup for the file, look up the given location, and return the definition path and location.

Possible future optimizations:
+ Avoid building the entire file's lookup table. Just build it for the given location. This would need getting the expression for a given location, which I don't think we have an API for right now. -- I ended up doing this later in the stack.

RFC:
I tried to come up with a test that would trigger the `ShadowedBy` branch, but could not. Will leave it as such unless others know how to trigger it.

Reviewed By: grievejia

Differential Revision: D34520603

fbshipit-source-id: 13090de7e8e85a3b77ad5e736e31a8321f88c075