Add backend variant for go-to-definition request.
Summary:
Note: I'm implementing the go-to-definition MVP with a straightforward query. Once we have tested it out and settled on the details needed from the backend, we can refactor the API as needed.

Note that this just takes in one path because there exists a `batch` query to batch multiple queries.

This diff adds the variant for the request.
Next diff: Add the response variant.

Reviewed By: grievejia

Differential Revision: D34391504

fbshipit-source-id: 59f1d669e400503b7dd4937196057b209c8a8baa