Add visitor to find the narrowest spanning expression for a position.
Summary:
Problem: The current implementation of `location_for_definition` uses the existing lookup table approach. However, that computes definitions (and resolved types) for all expressions in a file instead of just the one expression we care about. That is because we don't have a way to look up an expression given a position in a file.

This diff adds a visitor to find the narrowest expression that spans the given position.

This is largely adapted from the earlier visitor in the file `CreateDefinitionAndAnnotationLookupVisitor`, which does two things at once. Once we're happy with the new visitor, we can remove the definition-handling from that visitor. We can also optimize the `get_type` query in a similar fashion (later).

In case there are multiple expressions overlapping the position, I'm currently picking the first one. That seems to work well for the unit tests. The main times I've seen more than one element in the list is for spurious cases such as `foo()`, where qualification gets us `test.foo()`, which gives us two expressions with the same location range - `test.foo()` and `test`. Note that we don't handle `foo.bar()` at all right now. Can add logic to explicitly find the narrowest range later if this becomes an issue.

Reviewed By: grievejia

Differential Revision: D34520600

fbshipit-source-id: 03684c36ae0dbef2f596fa03c37d30558a672e4f