Don't compute definitions when computing the resolved type lookup table.
Summary:
Problem: We were doing both. This used to slow down both hover and go-to-def. Now that we've migrated to a fast, single-expression lookup for go-to-def, we can delete that part.

That slightly speeds up hover as well. Will optimize hover further and switch to an ondemand hover query later (once we are happy with the go-to-def strategy end-to-end).

Reviewed By: grievejia

Differential Revision: D34566036

fbshipit-source-id: e6f697b62aa649d7a11c5ef2af5df5f7f452242f