Clean up tests by having individual source strings - part 1.
Summary:
Problem: Each test was referring back to a long messy source file, which made it hard to know which position corresponded to which expression.

# Stack plan

The next few diffs remove the old-style definition lookup, which looked up definitions for all expressions.
+ Pro: That will decouple the type lookup and definition lookup.
+ Con: We might need it in case we hit some snag when testing the ondemand go-to-def end to end and decide to go the hover route.

Design choice: In case we decide to go the hover route, I can generalize the type lookup code instead of coupling them hard the way it is now. So, just deleting for now. RFC.

Reviewed By: grievejia

Differential Revision: D34566039

fbshipit-source-id: 29659238aecc99ed300f6ddd1e0ec1bdb58ec46d