Add configuration field for go-to-definition.
Summary:
Stack plan:
+ Add CLI flag and configuration field.
+ Add client-side data structures for the go-to-definition request and response.
+ Add a query for `location_for_definition(path, line, column)`. This will give us ondemand definition information for the MVP for go-to-def. Once we've settled on the implementation strategy for various features, we can refactor the query API to return more structured data as needed.

The part we are most uncertain about is whether users will have often wait for go-to-def respones because the server is busy updating on file changes. Only one way to find out! If Pyre is able to respond quickly even for large projects, then we can get away with an ondemand approach. Otherwise, we would have to cache it like we do for hover.

Next diff: Add CLI flag.

Reviewed By: grievejia

Differential Revision: D34391499

fbshipit-source-id: 5ffc60e94e5c30da8dfe1e8acca0098e987a9604