Query the server for location of definition.
Summary:
Problem: The client just returned a dummy response for go-to-def requests.

Now, we query the server and send the response.

For the MVP, I'm experimenting with an on-demand query. When we receive a go-to-def request, we will add it to the query queue. When the query comes up in the query manager loop, we will asynchronously query the server, receive the response, and send it to VS Code.

I'll check during dogfooding (along with devx help) if this takes too long. If so, we might need to go for the batch approach.

Note that we're not timing out if the server takes too long to respond. Might add that later.

Reviewed By: grievejia

Differential Revision: D34598992

fbshipit-source-id: b60869fbf8a34c033ebd2a24003afec0176dd23b