Remove sanitized flows when all sources or sinks have the same base
Summary:
When matching multiple sources with multiple sinks, we currently apply a small fixpoint where we sanitize sinks with the `NotSink[X]` taint transform from the sources, and then sanitize the sources with the `NotSource[X]` from the sinks, and repeat until we cannot remove anything.

Unfortunately, we are missing a special case where all sources have the same base (for instance: `Not[X]:A` and `Not[Y]:A` have the base `A`). In that case, it is safe to remove sinks that sanitize that source (for instance: in `X`, `Y`, `Not[A]:X` it is safe to remove `Not[A]:X`). This diff implements that logic.

See the integration test for a concrete example.

Reviewed By: dkgi

Differential Revision: D34633456

fbshipit-source-id: c2e8a3355ddee4f474e71f85713d7a4aeb240558