docs: add social banner in support of Ukraine (#595)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.
![image](https://user-images.githubusercontent.com/12485205/156675789-81bfb244-b27b-407f-9e20-86a66c7da8fb.png)

Pull Request resolved: https://github.com/facebook/pyre-check/pull/595

Test Plan:
Imported from GitHub, without a `Test Plan:` line.

**Static Docs Preview: pyre**
|[Full Site](https://our.intern.facebook.com/intern/staticdocs/eph/D34630286/V2/pyre/)|

|**Modified Pages**|

Reviewed By: 0xedward

Differential Revision: D34630286

Pulled By: dmitryvinn-fb

fbshipit-source-id: f1ea6f0f196b005924c06e48cba6f83879fda606