Support 'factory' in field descriptors
Summary:
# This diff

`factory` is an alias for `default_factory` in field descriptors for dataclass transforms.

# Context

Locally: `dataclasses` use the name `default_factory`, but `attrs` uses the name `factory` in many scenarios, so this alias was added for `dataclass_transforms` in the PEP.

See base diff and pointers there for broader context.

Reviewed By: grievejia

Differential Revision: D34487215

fbshipit-source-id: d628adfea5b9710158138b6b26d89cdc69653789