Add support for SanitizeSingleTrace
Summary:
We are still using the old sanitizers (`Sanitize` with `TaintSource` or `TaintSink`) in many places. These sanitizers only work in the forward or backward trace (not both), compared to the new sanitizers that are implemented using taint transforms.

We want to move away from these old sanitizers. The first step is to introduce a new syntax, `SanitizeSingleTrace`, which is equivalent to `Sanitize`, and will be used during the transition. This diff implements this.

In a follow-up, we will forbid the old sanitizer syntax, while still allowing `SanitizeSingleTrace`.

This diff also updates the documentation to use only the new sanitizer syntax. It documents `SanitizeSingleTrace` in the "advanced" section.
It also improves slightly error messages for syntax errors in sanitizers.

Reviewed By: dkgi

Differential Revision: D34664108

fbshipit-source-id: 1c25ecdda9e82038ca9093f9de6becbb394d2c60