Create intervals to represent subclasses
Summary:
The goal of this stack of diff is to filter out a particular class of traces (from sources to sinks). Such traces are featured by containing a consecutive sequence of `self` methods, whose invalidity is caused by not preserving the subclass information of the receiver of `self` methods. To efficiently represent the subclasses of a type, we can use intervals, which are thus called "class intervals". That is, the begin and end discovery time of a node in a class hierarchy tree (after DFSing the tree), which represents all of its children nodes. This diff implements such an interval data structure. Currently this is a wrapper around `Core.Interval.Int`, but later we may need to represent a class interval as a list of intervals, and need to support both intersection and union operations.

Additionally, this diff documents the traces that need to be removed, as integration tests.

More details can be found in the design doc.

Reviewed By: arthaud

Differential Revision: D34628225

fbshipit-source-id: b51406c4772c6fa36869d9bf691d18a425b66adf