Productionize incremental API (#596)
Summary:
Switch the API from using `uwsgi` over a socket to `gunicorn` over port 5000.

With this change, I am able to get it working in incremental mode with one pyre server per worker thread.

### Context for the changes

**Why the switch from `uwsgi` to `gunicorn`?**

They have different worker startup models: `uwsgi` does module initialization upfront and then forks, whereas gunicorn just starts multiple workers.

This is a problem for me because it is much easier to ensure each worker has an associated incremental server if we can force this initialization to happen at startup. There is a uwsgi decorator that should help with this, but I struggled to make it work properly, and the internet broadly suggests that gunicorn is a better choice anyway for simple webapps - uwsgi is more powerful but also quite a bit harder to fully understand because it is more complex.

**Why use port 5000 rather than a socket?**

A lot of guides use sockets, which in theory is simpler and more efficient. But if anything doesn't work, it puts us in a pickle because it's not straightforward to send HTTP requests to a socket, which means we can't easily determine whether a problem is with the server itself or the nginx config.

Running on port 5000 makes it easy to query the gunicorn server directly to determine that it is working, at which point any problems have to be with nginx.

It makes no difference from a security point of view given that port 5000 is closed (so you have to be ssh'ed in to hit it), and the overhead isn't a real concern.

For future reference I think it is possible to curl a socket, it's just quite a bit harder to remember the incantation, and the first few webpages I looked at gave me commands that didn't work. But here's a discussion of how to do it using the extra `--unix-socket`; the unintuitive thing is that the socket and the rest of the URL don't live in the same argument to the curl command
https://superuser.com/questions/834307/can-curl-send-requests-to-sockets

 ---

Pull Request resolved: https://github.com/facebook/pyre-check/pull/596

Test Plan:
The commands needed to spin up a playground server are unchanged, we're still using nginx plus a systemd service.

I rebooted the server running this code and saw that we now reliably finish typecheck requests in around a second.

Reviewed By: grievejia

Differential Revision: D34687333

Pulled By: stroxler

fbshipit-source-id: 28c918ded24a08e8ef71bc005c49d63aeb28bc13