Apply source and sink tito specific sanitizers on the return value for obscure models
Summary:
Marking the return value as taint-in-taint-out sanitized for a given source or sink should prevent taint from being propagated from any parameter to the return value. Unfortunately, that does not currently work as expected for obscure models, which leads to false positives.
This diff fixes that problem. See integration tests for examples.

Reviewed By: dkgi

Differential Revision: D34739272

fbshipit-source-id: c2234ab8e7ab416730ae826acc838672610f0023