Properly apply source and sink specific sanitizers on obscure models
Summary:
Marking a parameter as sanitized for a given source or sink should also prevent taint from being propagated from that parameter to the return value (also called taint-in-taint-out).

There is currently not properly applied for obscure models, which leads to false positives.
This diff fixes that problem. See integration tests for examples.

Reviewed By: alexkassil

Differential Revision: D34737438

fbshipit-source-id: 05d7487a54a098e62e0fffc0217ddf9d26bdc07d