Explicitly specifiy Shallow vs FineGrained in analysis style
Summary:
I was skimming the code to better understand exactly how environments are rebuilt,
and came across shallow vs fine-grained updates which may be relevant as we consider
our options for unsaved editor state.

It took more time to figure out that there are only two modes, Shallow and FineGrained,
because of the wildcard match statement; being explicit makes this more obvious (especially
since the explicit case was actually different in AstEnvironment vs IncrementalCheck, which
are the two places that care).

Reviewed By: grievejia

Differential Revision: D34774229

fbshipit-source-id: 3b8617b607b155ba36609d1428dcb5f84f71c1b5