Adjust decorator handling to have clear names and semantics
Summary:
These changes are debateable, but:
- we were wasting compute:
  - we were resolving the decorators many times in order to then compare
    to just one name, it makes much more sense to compare names in an
    inner loop
  - we computed a full list of decorator matches but never actually used
    it (except in a unit test) - 100% of list based uses were actually
    just calling is_empty
- the names were confusing; in particular get_decorator returned a whole
  list which is not what you'd expect - and we never actually used the
  full list except in a unit test. It's clearer and faster to just explicitly
  get the first match.

The performance benefits are nice, but my real motivation is just that I got
confused when trying to skim the code and answer these two questions:
- what does this function do?
- why would we ever call it?

Making it clearer means it's easier to reason about the environment stack

*NOTE*: I thought about making these functions more general by having them
just take decorators as an input rather than a class summary. But more general
isn't necessarily better - especially deep in the environment stack, it's helpful
to be able to guess what all of the uses of a function might be. So making the
functions specific to class decorators has two benefits:
- the callsites are all cleaner, since they don't have to reach down into the node
- it's easier to see the purpose of the function without having to look at all callers

Reviewed By: grievejia

Differential Revision: D34798477

fbshipit-source-id: 05cfd2cb23aa0c07dc729958fddf2c73731e7787