Do not distinguish between override calls and direct calls in the issue handle
Summary: Right now, we are using `Ovr{Class.method}` as the callee in the issue handle for calls that have overrides, and `Class.method` for direct calls. This can lead to issues being considered new when someone adds a child class of `Class`, even if it doesn't introduce new flows. Let's remove the `Ovr{}` from the handle so we do not trigger spurious new issues.

Reviewed By: alexkassil

Differential Revision: D34832215

fbshipit-source-id: 86a2cef5c30a07ce498fae664b083f8d595c5e29