Get rid of indirection
Summary:
I was briefly confused as to why we needed class_metadata_environment
to grab a location of a global. Solution: we don't, I have no idea
why this code ever existed.

Let's get rid of it; locations don't require anything but basic
AST lookup and our source code should reflect that. This also lets us
get rid of a duplicated chain of lookup functions; all that logic already
exists elsewhere if we just use it.

Reviewed By: shannonzhu

Differential Revision: D34829137

fbshipit-source-id: 11f303ccfe95ca7c45fa63a429458aa0ae5b0387