Support 3-arg version with literal attribute access
Summary:
We try to infer the type of `name` in a `getattr(object, name)` call. If the type of `name` is a constant literal string type `foo` (i.e. `Literal['foo']`) – i.e. we can statically conclude that `name` is always `"foo"` – we treat `getattr(object, name)` the same way as `object.foo`.

We weren't doing it for 3-arg version, leading to confusion for users, fix that.

Note that the type of default does not come into play when we know the attribute is definitely there.

There is another scenario where we somehow figure out that either the attribute is not there or if it is there it will always be of type `X`. Then, we could potentially a more precise type which is the join of `X` and type of default. That is outside the scope of this diff.

Reviewed By: grievejia

Differential Revision: D34774943

fbshipit-source-id: 19a606f9411dc128bd2cbf2f64477827049825c6