Rename SharedMemory.(TableType) value type t -> value
Summary:
For a Table module that is one of NoCache or WithCache, we
currently use Table.t to refer to the *value* type of a shared
memory table. This isn't great, because it's natural to think a
Table.t is "a table", in the sense that Map.t is "a map".

The problem is significantly compounded by the fact that so many
other modules inherit from one of these, and when skimming code it
can be tricky to remember this inheritance even if you've memorized
it at the base level.

It's also a problem when thinking about implementing overlays, because
we may wind up creating an actual value that controls how table access
works which would make it even more confusing for Table.t to mean the
value type rather than the table entity itself.

Reviewed By: shannonzhu

Differential Revision: D34901615

fbshipit-source-id: fb3193914b81f769b4fa2ffe4ff4f7bea33b97a0