Add pyre infer tests demonstrating greedy use of known types
Summary:
There was some confusion as to why pyre infer annotated the signature of a function that should take `Any` type with something more specific (D34812462). The example in the diff looked like:

```
def foo(x) -> Foo:
  if type(x) is Foo:
      return x
  else:
      cast_to_foo(x)
```

And pyre infer annotated `x` as having type `Foo`. It's pretty easy to see that even though `x` is clearly expected to be `Foo` some of the time, there are also other cases in the control flow where `x` can be something else that we don't know.

I'm labeling this as a "wont fix" for now despite the sometimes suboptimal inference outputs, because in a largely untyped context we stand to lose a lot if we throw away type information that we know verifiably describes a value in some cases if we don't know the type of the value in all cases.

Perhaps special casing refinement checks like `type(x) is Y` to avoid letting these provide any information to type inference is considerable, but then again runtime sanity checks on the type of a value are also not uncommon. If we invest more time later on to decreasing false positives in pyre infer, this is definitely something to look into. That said, I believe false negatives (inferrable types that we totally miss, or things we could take a reasonable guess at) are a bigger problem, and we don't even have the cycles to tackle that yet. So for now, let's just document this behavior with a test case.

Reviewed By: dkgi

Differential Revision: D34942678

fbshipit-source-id: fc832f6f2eb081250b8be84c2cc5d6ad5ecd7fb4