Don't barf on an alias that contains `"..."`.
Summary:
Problem: Our upgrade to the latest version of typeshed was blocked because of weird errors showing up (D34629788).

After substantial experimenting, I narrowed it down to `collections.abc` [re-exporting](https://github.com/python/typeshed/blame/f4ae363b564de69bfad4b4346204a25a99cd6642/stdlib/abc.pyi#L3) `Callable as Callable`. That was just the beginning, though. The next step was to figure out where the problem was - in the `Type.create` logic or somewhere in the unannotated global environment or alias environment logic.

Turns out that we handle the re-exported `Callable` just fine except when it is used with `...`. That makes us fail to recognize the alias.

Explanation:
+ We resolve an alias by finding its dependent aliases, resolving them, and then resolving the overall alias. So, for `F = TypeVar("F", bound=MyAlias)`, we see that `MyAlias` is a dependent, so we resolve that first, and then resolve the overall `F` `TypeVar` alias.
+ But when we had `F = TypeVar("F", bound=MyCallable[..., str])`, we would get dependents as (a) `MyCallable` and (b) `...`. And since `...` isn't an alias, we fail to resolve the whole thing.

Fix: Don't consider `...` to be a dependent alias.

In the end, it's a one-line change. Took me ~9 hours to debug and fix :P

justoneline

cc arthaud

Reviewed By: shannonzhu

Differential Revision: D34971834

fbshipit-source-id: 89b2b8bae528bc06d4e4cf806d7ff1f599e93e76