Avoid `open Memory` to improve searchability
Summary:
It is quite difficult to audit our code for the use of global,
mutable shared memory (whcih has to be done more-or-less perfectly
in order to correctly design support for unsaved changes - missing some
use case like Callgraph not explicitly made part of the environment
stack will lead to inconsistent state and very hard to track bugs).

Since we don't have great introspection tools for callgraphs, it's nice
if we can make simple code search as useful as possible for auditing.
The only use of `open Memory` in our project, and making all the Memory
uses explicit really doesn't hurt readability any.

Reviewed By: arthaud

Differential Revision: D34977277

fbshipit-source-id: e3ebe4448a92240fc0373c7be9369ea2a3076295