Do not duplicate interfaces in `hack_parallel`
Summary:
It looks to me like at one point in the past, there was the idea that
the `hack_parallel` std library might hide some implementation details
that were exposed in the interfaces of its component libraries. As
a result, hundreds of lines were copy-pasted verbatim.

In principle if this were very actively-maintained code it might be
worth keeping the hiding, but in practice:
- we can access member libraries more easily *directly* than via `Std`,
  which means there's effectively no hiding at all
- as a result we have hundreds of lines of junk to maintain, and in fact
  the interfaces have already gone out of sync, which confused me when I
  tried to simplify the `memory.ml` module
- this code is incredibly stable and not actively maintained, so the double
  interface is just dead weight, and easy to forget about

So I'm stripping it out - the interface is just the interface, no
double-interface with potential extra hiding.

There are exactly two differences introduced by this change based on a line-by-line
comparison:
- we now expose `Daemon.fork_FOR_TESTING_ON_UNIX_ONLY`. This seems fine, the name
  is crystal clear and as I mentioned above there's no real hiding anyway the
  way we use this code in pyre
- we expose `save_table` and `load_table` from `SharedMem`. This is actually
  desirable, we *intentionally* exposed them in D9436669 (https://github.com/facebook/pyre-check/commit/6a8634e94c6382512a02b714d09d9d5af6055589) but then forgot to
  update the second interface file - this kind of drift is exactly why the
  double interface isn't worth having if we don't really need it

Reviewed By: arthaud

Differential Revision: D34979624

fbshipit-source-id: 8690a3378303b959be46b3551746efd2a1a08b2c