Always use `SharedMemory`
Summary:
Presumably due to our no-abbreviations preference, we frequently
aliased the SharedMem hack_parallel module to SharedMemory.

This is not desirable, it makes skimming code and doing code
searches significantly harder if the module name can change
arbitrarily, our code style preferences are not a good enough
reason for this kind of confusion.

Reviewed By: grievejia

Differential Revision: D34979971

fbshipit-source-id: 529637ee3c062211d98032e02cb575474640727b