Move classInterval and tests into interprocedural/
Summary: Move files for computing class intervals into `source/interprocedural/`, such that we can compute the class intervals in `staticAnalysis.ml`. Otherwise, `analysisResult.mli` needs to depend on `classInterval.mli` and `classInterval.ml`, which are currently under `source/interprocedural_analyses/taint`. This is impossible because `analysisResult.mli` is a higher level module (because it is under `source/interprocedural`).

Reviewed By: arthaud

Differential Revision: D35020208

fbshipit-source-id: dc42c3284b59cdb191fa6660fc430b099ef4cb99