Gently compact the shared memory after global fixpoint iterations
Summary: Right now, we never compact the shared memory heap during the taint analysis. This leads to a very fragmented shared memory heap since every time we compute a new model, it will be written at the end of the heap. To save memory on the long term, let's compact the memory after each global fixpoint iteration. When the heap size is twice the used size, we will compact the heap. Note that the heap collection is pretty simple: all blocks are moved to the left (the optimal heap). Moving might be costly for big analyses.

Reviewed By: alexkassil

Differential Revision: D35033733

fbshipit-source-id: e48429bef36ffbb0780bd429efc6ad57dac05749