Wrap each error in a div
Summary:
Concatenating the errors with newlines does not work unless we're
in a `<pre>` tag. We could have used a pre tag, but that's not friendly
to later prettification (e.g. if we wanted to put some kind of red box
around each error), so instead I'm converting the errors to divs, which
is the more web-native way of putting them on separate lines.

Reviewed By: shannonzhu

Differential Revision: D35118982

fbshipit-source-id: df5f930a47ed98bbcb035622cc08a81b37744269