Allow setting code via a url parameter
Summary:
In the playground, historically the code editor has always been
initialized to `reveal_type(1)`.

This diff allows the user to set a `code` url parameter, and the
string value that is url-encoded will be used as the initial code.

This is the capability that we need to support having a sidebar
with examples for the PyCon tutorial.

I still need to work out how to set this url parameter when the user
clicks `Check`, which is what we'll need in order to make playground
state sharable; I'll tackle that in my next commit.

Reviewed By: dkgi

Differential Revision: D35119495

fbshipit-source-id: e90fadfa9187ed2bdb02e6f130251dc62d00f3e1