Set the url param when we submit
Summary:
In the previous commit, I modified the page so that when the `code`
url parameter is set, we'll pre-populate the code editor.

This commit makes the functionality much more useful by pushing
the url parameter into browser history when we click "submit",
so that it's easy to share links to examples.

Reviewed By: dkgi

Differential Revision: D35124782

fbshipit-source-id: 43d005a20ef88b41314f100bffdc7aab1951886f