Remove unecessary AstEnvironment.is_module
Summary:
This function is
(a) identical to `is_module_tracked`
(b) not used in any place where the alternate name is superior.

It's easier to have just one copy of the same function, and it's
also easier to use the same name as the ModuleTracker layer because
navigating all the layers is hard enough already and consistent
names reduces congitive load.

Reviewed By: arthaud

Differential Revision: D35221720

fbshipit-source-id: 3ddd821f83fdf94f06750599980a2ed0573ee9a2