Improve code editor sizing
Summary:
There's no one place to find instructions on how all the css
settings interact, but it appears that this combination of settings
allows the code area to be 25% of the full window height at a
minimum, 50% at a maximum, and automatically resize as needed.

We could adjust these settings if we want - e.g. if we want to
prioritize behavior on a super zoomed-in view we might want to
set min-height in absolute numbers rather than as a percent of
viewport height.

Reviewed By: shannonzhu

Differential Revision: D35259820

fbshipit-source-id: def2214816e47a1f5c25e340ec3e28d2aae0c1db