ModuleTracker.ReadOnly [2]: remove unused lookup logic
Summary:
While working on a moduleTracker.ReadOnly interface, I
realized that the `lookup` function and associated output datatype
are not used (other than one unit test).

Since implementing `lookup` leads to redundant logic, I'd like
to propose getting rid of it - this leads to a slimmer and easier
to skim interface, and less code to maintain.

Reviewed By: grievejia

Differential Revision: D35228181

fbshipit-source-id: 82ac2d55f7c4ab99d72d75eeb5845091c1e48557