Remove explicit_module_counts function
Summary:
This function was used in two tests and in one logging statement; the
only benefit of having it rather than just using `List.length source_paths`
is that it's more efficient, but neither of the two use cases are
in a place where performance matters.

Removing it simplifies the interface and makes it clearer what operations
are really essential in ModuleTracker.

Reviewed By: grievejia

Differential Revision: D35234050

fbshipit-source-id: 3ce9bb71e7aa83667863f7909df7fd388c17613a