ModuleTracker.ReadOnly [1]: Extract read-only interface
Summary:
See previous commit for rationale on why to create a read-only
ModuleTracker interface; the short version is that it will allow us
to abstract the filesystem, which is needed both to improve our tests
and to enable overlays for unsaved changes.

This commit focuses on updating the API and all callsites without
modifying the implementation (for now a `ModuleTracker.ReadOnly.t`
is the same record as a `ReadOnly.t`); in the next commit I'll
change the internal representation to use callbacks.

Reviewed By: grievejia

Differential Revision: D35228179

fbshipit-source-id: a3e07363e4533336e93c8e50ad263faa7104d37a