ModuleTracker.ReadOnly [0]: reorganize code
Summary:
As part of implementing support for unsaved changes, I intend to create
a read-only ModuleTracker API that will allow us to abstract all interactions
with the filesystem in terms of callbacks to a read-only API (similar to how
all our other read-only apis are implemented via records of callbacks).

Part of this refactor will also include pushing the raw source reading down
into ModuleTracker from AstEnvironment, but that change is not in this diff.

By doing this, we will be able to:
- easily handle the overlay of unsaved changes on top of the filesystem
  by intercepting source code lookups
- moreover, we can make the ScratchProject module tracker use a customized
  implementation that doesn't touch the filesystem; this ought to make
  our tests faster and change the flakey CI results we get right now
  due to filesystem operations in unit tests

This specific diff is small, I'm just moving the operations that need
to exist in our read-only API down to the bottom so that I can reorganize
them later.

Reviewed By: grievejia

Differential Revision: D35228178

fbshipit-source-id: 7831de1dc62281bd7fad4a682b1652aa3a41d64b