Use the return type annotation on the callee when possible
Summary:
## Problem
Resolving a full call expression can be very costly when the callee is a union of a big number of elements. This leads to our call graph construction being very slow (sometimes, as slow as the full taint analysis).

## Solution
When possible, we should use the return type annotation of the callee. In general, this is actually beneficial since we can infer different return types for different callees. In that diff, we perform this logic and we store the return type for each different target (i.e, `CallTarget`) instead of the whole call site (i.e, `CallCallees`). For cases where inferring from the return type annotation would resolve into a less precise type (in the presence of oveloads or type variables), we fall back to resolving the full call expression, as before. See the integration tests for examples.

Differential Revision: D35333447

fbshipit-source-id: 0e01a99f64a05f9c4c70ec57715c156129d29d85