We don't need configuration for dependency transactions.
Summary:
After some discussion Friday, we've tentatively decided we might
want to refactor the code so that ModuleTracker.t contains a copy
of the configuration, which would save us from needing to pass it
around explicitly - this makes pretty good sense since
- a configuration is needed to build a module tracker in the first place
- if the configuration changes we always restart, so the environment stack is tightly coupled to the configuration and cannot outlive it - and therefore might as well contain it

But as I was looking into what this would take and what the effect would
be, I noticed that we need configuration to run updates. Part of the reason
for this is shallow vs fine-grained dependencies, which I think we should
try to refactor out of the configuration. But the other part is that
DepedencyTrackedMemory requires a configuration for transactions.

This puzzled me - why would we need a configuration for simple dependency
tracking operations? The answer is we don't; my guess is that maybe at one
point the dependency tracking was an experimental feature hidden behind a
flag, but today the configuration is just there in the signatures and
completely unused. So let's get rid of it!

Reviewed By: grievejia

Differential Revision: D35361219

fbshipit-source-id: dde21f578596256aafde940f1a404a51aae1ec21